requirements.txt
PySocks==1.7.1
python-slugify==8.0.4
pytools==2024.1.11
PyYAML==6.0.1
safetensors==0.4.3
scipy==1.10.1
seaborn==0.13.2
sentry-sdk==2.12.0
setproctitle==1.3.3
shortuuid==1.0.13
smmap==5.0.1
sympy==1.12
text-unidecode==1.3
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchprofile==0.0.4
torchvision==0.19.0
tqdm==4.66.4
transformers==4.43.3
typing_extensions==4.9.0
ultralytics==8.2.68
ultralytics-thop==2.0.0
matplotlib==3.7.5
numpy==1.24.1
pandas==2.0.3
pillow==10.2.0
protobuf==5.27.2
pynvml==11.5.3
pyopencl==2024.2.7
sentencepiece==0.2.0
thop==0.1.1.post2209072238
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
huggingface-hub==0.24.2
kiwisolver==1.4.5
mpmath==1.3.0
networkx==3.0
opencv-python==4.10.0.84
ftfy==6.2.3

